Fixed charge behavior

+++ lib/Core/Behaviors/AttackBehaviors/Attack/ChargeAttack/ChargeAttack.cpp
@@ -8,7 +8,7 @@ ChargeAttack::ChargeAttack(IMotorController* motorController) :
 
 ChargeAttack::Status ChargeAttack::execute(const SensorData data) {
 
-    float diff = static_cast<int>((data.right - data.left) * (1 / 1000.0));
+    float diff = (data.right - data.left) * (1 / 1000.0);
 
     m_motorController->sameDirection(255, diff);
